fix: remove files-copy directory each time

+++ 04-copy-directory/index.js
@@ -1,27 +1,36 @@
 const fs = require("fs/promises");
 const path = require("path");
 
-
 async function copyDirectory() {
   try {
-    const projectDistPath = path.join(__dirname, "files-copy");
-    await fs.mkdir(projectDistPath, { recursive: true });
+    const sourceDir = path.join(__dirname, "files");
+    const destinationDir = path.join(__dirname, "files-copy");
+
+    // Создание папки назначения, если она отсутствует
+    await fs.mkdir(destinationDir, { recursive: true });
     console.log("Folder is created");
 
-    const sourceDir = "./files";
-    const destinationDir = "./files-copy";
-    const files = await fs.readdir(sourceDir, { withFileTypes: true });
-
-    for (const file of files) {
-      const sourcePath = path.join(sourceDir, file.name);
-      const destinationPath = path.join(destinationDir, file.name);
-      await fs.copyFile(sourcePath, destinationPath);
+    // Очистка папки назначения
+    const existingFiles = await fs.readdir(destinationDir);
+    for (const file of existingFiles) {
+      await fs.unlink(path.join(destinationDir, file));
     }
-    console.log('Success!');
-  } catch(err) {
-    console.log("Err", err);
-  }
 
+    // Копирование файлов из sourceDir в destinationDir
+    const files = await fs.readdir(sourceDir, { withFileTypes: true });
+    for (const file of files) {
+      if (file.isFile()) {
+        const sourcePath = path.join(sourceDir, file.name);
+        const destinationPath = path.join(destinationDir, file.name);
+        await fs.copyFile(sourcePath, destinationPath);
+      }
+    }
+
+    console.log("Success!");
+  } catch (err) {
+    console.log("Error:", err);
+  }
 }
 
 copyDirectory();
+
